Update thermocouples.h
change float type to double for "get_temp_celsius"

+++ src/thermocouples.h
@@ -10,6 +10,6 @@ namespace thermocouples
     void read_temp_from_one_thermocouple_once(thermocouple_index);
     void read_temp_from_one_thermocouple(thermocouple_index);
     void stop_reading_temp();
-    float get_temp_celsius(thermocouple_index);
+    double get_temp_celsius(thermocouple_index);
 }
